PP-3714 - Better validation
Client side field validation on DD didn't allow for custom error
messages on from the validator function, this means that if you're
validating a field against more than one validator you can only show one
error message, this is sad. So now the validator passes back a message
which is used as the error message, if it’s not present it defaults to
the data attribute error message that is set.