PP-3838 Add mandate state enforcer
This commit adds mandate state enforcer middleware.

It ensures that the user's place in the flow matches the
current mandate state, and if not shows an appropriate error
page. We have chosen to only look at mandate state for now,
as this is what we actually care about. It could be extended to payment
request for one off payments if need be.

We have chosen not to throw an error if a mandate does not exist, as this
is the job of the `getMandate` middleware.

We have not exhaustively tested all error messaging, just the mechanism
for showing the correct message for a given payment state.