PP-5533 - make button bigger on focus to catch clicks
Because we have JS validating on blur there is
an edge case (detailed in PP-5533) where when clicking the
submit button when you‘re focused on a errored
field causes the page to jump up and then the
click event misses the button.

So we can fix this by making the button invisibly
bigger when it becomes active.

Hat tip for @tombye for the solution ✊