PP-5726 Integrate app with Sentry
Adding sentry with minimal disruption to
current error handling.

Add sentry to server.js.
Add default request handler and error handler
as first (and currently only) error middleware.

Use renderErrorView as central point to emit
events to Sentry.
In order to do this we pass the original error
in where available. Only send errors to sentry when
status code is 500 or greater, since we don't want
to be alerted on client errors.

with @mrlumbu